Plant Disease Detection/Api/requirements.txt
tensorflow==2.16.1
fastapi
uvicorn
python-multipart
pillow
tensorflow-serving-api==2.16.1
matplotlib
numpy

